+++ src/commands/outputs/get-item.ts
@@ -1,8 +1,8 @@
 import type { GetCommandOutput } from "@aws-sdk/lib-dynamodb";
 import { trusted } from "@infra-blocks/types";
 import type { Attributes } from "../../types.js";
+import { ifDefined, unsetUndefined } from "../lib.js";
 import { ConsumedCapacity } from "./consumed-capacity.js";
-import { mapIfDefined, unsetUndefined } from "./lib.js";
 
 export type GetItemOutput<T extends Attributes = Attributes> = {
   item?: T;
@@ -17,8 +17,8 @@ function decode<T extends Attributes = Attributes>(
   output: GetCommandOutput,
 ): GetItemOutput<T> {
   return unsetUndefined({
-    item: mapIfDefined(output.Item, trusted<T>),
-    consumedCapacity: mapIfDefined(
+    item: ifDefined(output.Item, trusted<T>),
+    consumedCapacity: ifDefined(
       output.ConsumedCapacity,
       ConsumedCapacity.decode,
     ),