+++ src/commands/outputs/update-item.ts
@@ -1,7 +1,7 @@
 import type { UpdateCommandOutput } from "@aws-sdk/lib-dynamodb";
 import { trusted } from "@infra-blocks/types";
 import type { Attributes } from "../../types.js";
-import { mapIfDefined, unsetUndefined } from "./lib.js";
+import { ifDefined, unsetUndefined } from "../lib.js";
 
 export type UpdateItemOutput<T extends Attributes = Attributes> = {
   attributes?: T;
@@ -15,6 +15,6 @@ function decode<T extends Attributes = Attributes>(
   output: UpdateCommandOutput,
 ): UpdateItemOutput<T> {
   return unsetUndefined({
-    attributes: mapIfDefined(output.Attributes, trusted<T>),
+    attributes: ifDefined(output.Attributes, trusted<T>),
   });
 }