+++ test/unit/commands/outputs/index.ts
@@ -1,10 +1,12 @@
 import { suite } from "node:test";
 import { consumedCapacityTests } from "./consumed-capacity.js";
+import { deleteItemTests } from "./delete-item.js";
 import { getItemTests } from "./get-item.js";
 
 export const outputsTests = () => {
   suite("outputs", () => {
     consumedCapacityTests();
+    deleteItemTests();
     getItemTests();
   });
 };