+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.72.0] - 2026-03-07
+
+### Added
+
+- Support for `returnConsumedCapacity` for `DeleteItem`.
+
 ## [0.71.0] - 2026-03-05
 
 ### Added
@@ -36,7 +42,7 @@ should be. The fix resolves to `Value<string | number> | number`, which accepts:
 
 ### Added
 
-- Support for `returnConsumedMetrics` for `GetItem`.
+- Support for `returnConsumedCapacity` for `GetItem`.
 
 ### Changed
 
@@ -913,6 +919,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.72.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.71.0...v0.72.0
 [0.71.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.70.1...v0.71.0
 [0.70.1]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.70.0...v0.70.1
 [0.70.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.69.0...v0.70.0